Refactored the async cell image load to use cancelable NSOperations
This will allow us to tie one NSOperation block to each cell. Then
we can control canceling previous operations that were still pending
when the cell is going to be re-used.  If we can't cancel the previous
operations for a cell we can end up showing the wrong image for the
cell.